feat: merge plugin not ending in plugin
