--- bevy_lint/tests/ui/plugin_not_ending_in_plugin/main.stderr
@@ -1,21 +1,0 @@
-error: implemented `Plugin` for a structure whose name does not end in "Plugin"
-  --> tests/ui/plugin_not_ending_in_plugin/main.rs:9:8
-   |
-9  | struct Foo;
-   |        ^^^ help: rename the plugin: `FooPlugin`
-   |
-note: `Plugin` implemented here
-  --> tests/ui/plugin_not_ending_in_plugin/main.rs:14:1
-   |
-14 | / impl Plugin for Foo {
-15 | |     fn build(&self, _app: &mut App) {}
-16 | | }
-   | |_^
-note: the lint level is defined here
-  --> tests/ui/plugin_not_ending_in_plugin/main.rs:3:9
-   |
-3  | #![deny(bevy::plugin_not_ending_in_plugin)]
-   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-error: aborting due to 1 previous error
-
