+++ bevy_lint/tests/ui/unconventional_naming/main.stderr
@@ -16,5 +16,19 @@ note: the lint level is defined here
    |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^
    = note: this error originates in the derive macro `SystemSet` (in Nightly builds, run with -Z macro-backtrace for more info)
 
-error: aborting due to 1 previous error
+error: unconventional type name for a `Plugin` or `SystemSet`
+  --> tests/ui/unconventional_naming/main.rs:15:8
+   |
+15 | struct Foo;
+   |        ^^^ help: rename the Plugin: `FooPlugin`
+   |
+note: `Plugin` implemented here
+  --> tests/ui/unconventional_naming/main.rs:20:1
+   |
+20 | / impl Plugin for Foo {
+21 | |     fn build(&self, _app: &mut App) {}
+22 | | }
+   | |_^
+
+error: aborting due to 2 previous errors
