+++ bevy_lint/tests/ui/unconventional_naming/spoofed_name.stderr
@@ -1,0 +1,21 @@
+error: unconventional type name for a `Plugin` or `SystemSet`
+  --> tests/ui/unconventional_naming/spoofed_name.rs:13:20
+   |
+13 |         pub struct Foo;
+   |                    ^^^ help: rename the Plugin: `FooPlugin`
+   |
+note: `Plugin` implemented here
+  --> tests/ui/unconventional_naming/spoofed_name.rs:23:1
+   |
+23 | / impl Plugin for FooPlugin {
+24 | |     fn build(&self, _app: &mut App) {}
+25 | | }
+   | |_^
+note: the lint level is defined here
+  --> tests/ui/unconventional_naming/spoofed_name.rs:6:9
+   |
+6  | #![deny(bevy::unconventional_naming)]
+   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+error: aborting due to 1 previous error
+
